Move QueryArgs to own class
